Move BI specific icons to cmk.gui.bi
This prevents an import cycle caused in the UI code.
