+++ cmk/gui/bi/_icons.py
@@ -1,0 +1,112 @@
+#!/usr/bin/env python3
+# Copyright (C) 2019 Checkmk GmbH - License: GNU General Public License v2
+# This file is part of Checkmk (https://checkmk.com). It is subject to the terms and
+# conditions defined in the file COPYING, which is part of this source code package.
+
+from cmk.utils.tags import TagID
+
+from cmk.gui.http import request
+from cmk.gui.i18n import _
+from cmk.gui.logged_in import user
+from cmk.gui.type_defs import Icon as IconSpec
+from cmk.gui.type_defs import Row
+from cmk.gui.utils.html import HTML
+from cmk.gui.utils.urls import makeuri_contextless, urlencode
+from cmk.gui.views.icon import Icon, IconRegistry
+
+from ._compiler import is_part_of_aggregation
+
+
+def register(
+    icon_and_action_registry: IconRegistry,
+) -> None:
+    icon_and_action_registry.register(AggregationsIcon)
+    icon_and_action_registry.register(AggregationIcon)
+
+
+class AggregationsIcon(Icon):
+    @classmethod
+    def ident(cls):
+        return "aggregations"
+
+    @classmethod
+    def title(cls) -> str:
+        return _("Aggregations")
+
+    def render(
+        self,
+        what: str,
+        row: Row,
+        tags: list[TagID],
+        custom_vars: dict[str, str],
+    ) -> None | IconSpec | HTML | tuple[IconSpec, str] | tuple[IconSpec, str, str]:
+        # Link to aggregations of the host/service
+        # When precompile on demand is enabled, this icon is displayed for all hosts/services
+        # otherwise only for the hosts/services which are part of aggregations.
+        if is_part_of_aggregation(row["host_name"], str(row.get("service_description"))):
+            view_name = "aggr_%s" % what
+
+            if not user.may("view.%s" % view_name):
+                return None
+
+            urivars = [
+                ("view_name", view_name),
+                ("aggr_%s_site" % what, row["site"]),
+                ("aggr_%s_host" % what, row["host_name"]),
+            ]
+            if what == "service":
+                urivars += [("aggr_service_service", row["service_description"])]
+            url = makeuri_contextless(request, urivars, filename="view.py")
+            return (
+                "aggr",
+                _("BI Aggregations containing this %s")
+                % (what == "host" and _("Host") or _("Service")),
+                url,
+            )
+        return None
+
+
+class AggregationIcon(Icon):
+    @classmethod
+    def ident(cls):
+        return "aggregation_checks"
+
+    @classmethod
+    def title(cls) -> str:
+        return _("Aggregation checks")
+
+    def host_columns(self):
+        return ["check_command", "name", "address"]
+
+    def render(
+        self,
+        what: str,
+        row: Row,
+        tags: list[TagID],
+        custom_vars: dict[str, str],
+    ) -> None | IconSpec | HTML | tuple[IconSpec, str] | tuple[IconSpec, str, str]:
+        # service_check_command looks like:
+        # u"check_mk_active-bi_aggr!... '-b' 'http://localhost/$HOSTNAME$' ... '-a' 'Host foobar' ..."
+        if what == "service" and row.get("service_check_command", "").startswith(
+            "check_mk_active-bi_aggr!"
+        ):
+            args = row["service_check_command"]
+            start = args.find("-b' '") + 5
+            end = args.find("' ", start)
+            base_url = args[start:end].rstrip("/")
+            base_url = base_url.replace("$HOSTADDRESS$", row["host_address"])
+            base_url = base_url.replace("$HOSTNAME$", row["host_name"])
+
+            start = args.find("-a' '") + 5
+            end = args.find("' ", start)
+            aggr_name = args[start:end]
+            aggr_name = aggr_name.replace("$HOSTADDRESS$", row["host_address"])
+            aggr_name = aggr_name.replace("$HOSTNAME$", row["host_name"])
+
+            url = "{}/check_mk/view.py?view_name=aggr_single&aggr_name={}".format(
+                base_url,
+                urlencode(aggr_name),
+            )
+
+            return "aggr", _("Open this Aggregation"), url
+        return None
