+++ cmk/gui/plugins/main_modules/registration.py
@@ -177,6 +177,7 @@ def register() -> None:
         main_module_topic_registry,
         main_module_registry,
         mode_registry,
+        icon_and_action_registry,
     )
     robotmk.register(page_registry)
     cron.register(page_registry)
